# openapi.yaml
openapi: 3.0.1
info:
  title: TrainMe API
  version: v0.0.1
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /api/user/:
    get:
      security:
        - Token: []
      tags:
        - User API
      operationId: getCurrentUser
      responses:
        "200":
          description: Get current logged user information
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/UserDTO"
        "401":
          description: "Error: unauthorized"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Unauthorized"
        "500":
          description: "Internal server error, when error on server side"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/InternalServerError"
        "400":
          description: "Bad request error"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/BadRequestError"
    put:
      security:
        - Token: []
      tags:
        - User API
      operationId: updateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserDTO"
        required: true
      responses:
        "200":
          description: Change current logged user information
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: "Error: unauthorized"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Unauthorized"
        "500":
          description: "Internal server error, when error on server side"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/InternalServerError"
        "400":
          description: "Bad request error"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/BadRequestError"
...
  /api/user/{userId}:
    get:
      security:
        - Token: []
      tags:
        - User API
      operationId: getUserProfile
      parameters:
        - name: userId
          in: path
          required: true
          description: Numeric id of the user to get
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Get user profile by id (only for admins)
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/UserDTO"
        "401":
          description: "Error: unauthorized"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Unauthorized"
        "500":
          description: "Internal server error, when error on server side"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/InternalServerError"
        "403":
          description: "Forbidden response for no-admin user"
        "400":
          description: "Bad request error"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/BadRequestError"
...
components:
  securitySchemes:
    Token:
      type: apiKey
      in: header
      name: Authorization
  schemas:
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        username:
          type: string
          example: "testusername"
        dateOfBirthday:
          type: string
          format: date
          example: "2022-12-12"
        membershipFrom:
          type: string
          format: date
          example: "2022-12-12"
        membershipTo:
          type: string
          format: date
          example: "2022-12-12"
        memberships:
          uniqueItems: true
          type: array
          items:
            type: string
            enum:
              - ROLE_STANDARD
              - ROLE_ADMIN
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: "successfully"
...
    Unauthorized:
      type: object
      properties:
        username:
          type: string
          example: "Invalid Username"
        password:
          type: string
          example: "Invalid Password"
    InternalServerError:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2023-01-04T16:46:17.825+00:00"
        status:
          type: integer
          format: int64
          example: 500
        error:
          type: string
          example: "Internal Server Error"
        path:
          type: string
          example: "error path"
    BadRequestError:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2023-01-04T16:46:17.825+00:00"
        status:
          type: integer
          format: int64
          example: 400
        error:
          type: string
          example: "Bad Request"
        exception:
          type: string
          example: "exception method on server"
        message:
          type: string
          example: "error message"
        path:
          type: string
          example: "error path"